Mediate touch on dependents
In Rails 2.3, touch was implemented with an ultimate save!, and would
therefore be mediated so long as the save methods were.  In Rails 3.x,
touch is implemented with direct sql and needs to be mediated in
dependent classes.  This fixes the failure of
spec/integration/locking_tests_spec.rb:87 'should increment lock_version
for the graph if a dependent is changed'.