Update locking test for dependents with lock_version
With the change to touch being a direct sql call in Rails 3.x, we need
to use update_attributes on a stale party to trigger the
StaleObjectError caused by dependents with lock_version having been
updated by their dependents counter_caches.

Also fixed a problem in the reservations test model with party_lodgings
creation not referencing the passed lodging correctly.